move span into lib
